CodeAlpha_Handwritten_Character_Recognition/requirements.txt
torch
torchvision
numpy
pandas
matplotlib
opencv-python
scikit-learn
Pillow
